pyproject.toml
[project]
name = "ravi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "dotenv>=0.9.9",
    "googlesearch-python>=1.3.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.27",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "streamlit>=1.44.1",
]
